requirements.txt
discord.py[voice]
yt-dlp
PyNaCl
flask
python-dotenv
imageio-ffmpeg